[v2.11.0]Beta

+++ cogs/report/report_send_files.py
@@ -26,12 +26,9 @@ class ReportSendFiles(commands.Cog):
     if custom_id == "report_send_file":
       path = f"data/report/private_report/{interaction.guild.id}.json"
       if not os.path.exists(path):
-        e = f"[ERROR[3-5-01]]{datetime.datetime.now()}\n- GUILD_ID:{interaction.guild.id}\nJson file was not found"
+        e = f"[ERROR[3-2-01]]{datetime.datetime.now()}\n- GUILD_ID:{interaction.guild.id}\nJson file was not found"
         print(e)
-        embed=error.generate(
-          code="3-5-01",
-          description="サーバーデータが存在しませんでした。\nサポートサーバーまでお問い合わせください。"
-        )
+        embed=await error.generate("3-2-01")
         await interaction.response.send_message(embed=embed, ephemeral=True)
         return
 
@@ -91,16 +88,13 @@ class ReportSendFiles(commands.Cog):
       async with aiofiles.open(path, encoding='utf-8', mode="r") as f:
         contents = await f.read()
       report_dict = json.loads(contents)
+      user_id = report_dict.get(str(interaction.channel.id))
 
-      try:
-        user_id = report_dict[str(interaction.channel.id)]
-      except KeyError:
-        e = f"\n[ERROR[3-5-02]]{datetime.datetime.now()}\n- GUILD_ID:{interaction.guild.id}\n- CHANNEL_ID:{interaction.channel.id}\nReporter_id was not found\n"
+      # report者がNoneの場合->return
+      if not user_id:
+        e = f"\n[ERROR[3-2-02]]{datetime.datetime.now()}\n- GUILD_ID:{interaction.guild.id}\n- CHANNEL_ID:{interaction.channel.id}\nReporter_id was not found\n"
         print(e)
-        embed=error.generate(
-          code="3-5-02",
-          description="ユーザーデータが存在しませんでした。\nサポートサーバーまでお問い合わせください。"
-        )
+        embed=await error.generate(code="3-2-02")
         await interaction.followup.send(embed=embed)
         await interaction.message.delete()
         await self.add_reply(interaction)
@@ -109,10 +103,7 @@ class ReportSendFiles(commands.Cog):
       try:
         user = await interaction.guild.fetch_member(user_id)
       except Exception:
-        embed=error.generate(
-          code="3-5-03",
-          description="匿名Reportのユーザーを取得することができませんでした。\nユーザーは既にサーバーを抜けているかも...？"
-        )
+        embed=await error.generate(code="3-2-03")
         await interaction.followup.send(embed=embed, ephemeral=True)
         await interaction.message.delete()
         await self.add_reply(interaction)
@@ -121,13 +112,13 @@ class ReportSendFiles(commands.Cog):
       # embedを定義
       embed = discord.Embed(
         url = interaction.channel.jump_url,
-        description="## 匿名報告\n"
+        description="## 匿名Report\n"
                     f"あなたの報告に、『{interaction.guild.name}』の管理者からファイルが届きました。\n"
                     f"- __**このメッセージに返信**__(右クリック→返信)すると、このサーバーの管理者に届きます。",
         color=0xF4BD44,
       )
       embed.set_footer(
-        text=f"匿名報告 | {interaction.guild.name}",
+        text=f"匿名Report | {interaction.guild.name}",
         icon_url=interaction.guild.icon.replace(format='png').url if interaction.guild.icon else None,
       )
 
@@ -135,12 +126,9 @@ class ReportSendFiles(commands.Cog):
       try:
         files = [await attachment.to_file() for attachment in message.attachments]
       except Exception as e:
-        e = f"\n[ERROR[3-5-04]]{datetime.datetime.now()}\n- GUILD_ID:{interaction.guild.id}\n{e}\n"
+        e = f"\n[ERROR[3-2-04]]{datetime.datetime.now()}\n- GUILD_ID:{interaction.guild.id}\n{e}\n"
         print(e)
-        embed=error.generate(
-          code="3-5-04",
-          description="ファイル変換時に、不明なエラーが発生しました。\nサポートサーバーまでお問い合わせください。",
-        )
+        embed=await error.generate(code="3-2-04")
         await interaction.followup.send(embed=embed, ephemeral=True)
         await interaction.message.delete()
         await self.add_reply(interaction)
@@ -150,20 +138,14 @@ class ReportSendFiles(commands.Cog):
       try:
         await user.send(embed=embed, files=files)
       except discord.errors.Forbidden:
-        embed=error.generate(
-          code="3-5-05",
-          description="匿名Report送信者がDMを受け付けてないため、送信されませんでした。",
-        )
+        embed=await error.generate(code="3-2-05")
         await interaction.followup.send(embed=embed)
         await interaction.message.delete()
         return
       except Exception as e:
-        e = f"\n[ERROR[3-5-06]]{datetime.datetime.now()}\n- GUILD_ID:{interaction.guild.id}\n{e}\n"
+        e = f"\n[ERROR[3-2-06]]{datetime.datetime.now()}\n- GUILD_ID:{interaction.guild.id}\n{e}\n"
         print(e)
-        embed=error.generate(
-          code="3-5-06",
-          description="不明なエラーが発生しました。サポートサーバーまでお問い合わせください。",
-        )
+        embed=await error.generate(code="3-2-06")
         await interaction.followup.send(embed=embed)
         await interaction.message.delete()
         await self.add_reply(interaction)
